chore: Revise type hinting and object instantiation in finance service
Improved type hinting and object instantiation in the "AbstractApplicationService" and "AddLoanApplicationCommandHandler" classes. The application now uses more specific class objects, like using "RepositoryContextSession" for context in the abstract service, and directly initializing "LoanRepository" in the loan service. The changes aim to enhance code readability and maintain consistency across the application.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>