refactor: Generate loan ID in application handler and adjust loan repository
UUID generation for loan IDs has been moved from the repository to the add_loan_application_handler. Alongside this, updates to the 'add' method in the loan_repository has been made to accept the LoanRequest object instead of a generic request and to return an AddLoanApplicationResponseDTO. These changes improve data consistency and ensure a separation of responsibilities.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>