refactor: Refactor code to remove User model and improve repository usage
This commit removes the User model across the application which includes classes and tests. It also updates the repositories to allow direct retrieval of loan and loan account data. The LoanAccount and LoanRequest models have been updated to reflect these changes. The loan services and DTOs are likewise updated to work with these changes in models and repositories.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>