chore: Update layout and data within the LoanAccountComponent
This commit makes several changes in the LoanAccountComponent. It refactors the layout to display the standard interest rate by adjusting the span text. It also adds a button labeled "Reecheble Loans" and adjusts some data attributes within the JSON file to include "interest_rate". Lastly, it modifies the table headers in the LoanAccount component to change the sortable columns, including the change from 'name' to 'reference'.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>