chore: Implement unique loan reference for loan accounts
Introduced a new 'reference' field to the LoanRequest model, which creates a unique reference for each loan in the system. This reference is generated through create_loan_reference() static method. All services and repositories have been updated to make use of this reference instead of the previously used UUID for identifying loans. Entities such as AddLoanApplicationResponseDTO and PayLoanInstallmentResponseDTO were updated to return the new loan reference.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>