fix: Correct field names in loan component
This commit updates several incorrect and inaccurate field names in the loan component's HTML and data JSON. This includes changing 'price' to 'amount', 'category' to 'interestRate' and 'duration', 'rating' to 'originationDate' and 'dueDate', and finally 'price' to 'equatedMonthlyInstallment'. Updating these names improves readability and understanding for future code maintainability. Also, renamed "interest_rate" to "interestRate" in JSON data files for

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>