chore: Update p-table dataKey in loan component
Adjusted the dataKey properties in p-table elements within the loan component. The main table's identifying key was changed from 'name' to 'reference'. In the expanded row table, the key was modified from 'id' to 'month', aligning with changes in the data model. The naming for individual loan instances within the expanded row template was also corrected from 'loans' to 'loan'.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>