refactor: Refactor code for better representation of loans
Refactored various parts of the code to better represent loans instead of products. This included changing variable names in the loan account component's HTML and updating the import paths for the Loan model in several files. The Loan and RepaymentHistory interfaces were also converted to classes for better data encapsulation.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>