refactor: Refactor loan features and add LoanService
The code was significantly refactored, mainly focusing on loan features. Update involved a change in 'loan-account.component.html' where 'accountId' was changed to 'accountNumber'. Additionally, unnecessary columns of the loans table were removed, and 'loan-model.ts' was greatly expanded, with new classes and methods related to the loan functionality. A new 'LoanService' was introduced, allowing for better management of the loan operations, including fetching account loans. This refactor provides better clarity and organization for the codebase.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>