refactor: Refactor LoanAccountModel class to LoanAccountApiResponseModelData
The class LoanAccountModel in the loan-account-model has been renamed to LoanAccountApiResponseModelData. This change affects various areas of the code where LoanAccountModel was previously used, and was replaced accordingly. Such changes apply throughout the LoanAccountService, and the loan-accounts component that utilise this model, ensuring the appropriate refactoring, adjustments, and replacements are made throughout, providing naming consistency.

Signed-off-by: MliloM <mduduzimlilo@hotmail.com>